requirements.txt
streamlit==1.38.0
python-docx==1.1.2
openai==1.47.0
python-dotenv==1.0.1